backend/requirements.txt
Django==3.2.6
djoser==2.1.0
flake8==5.0.4
gunicorn==20.1.0
isort==5.10.1
python-dotenv==1.0.0
djangorestframework==3.12.4
webcolors==1.11.1
Pillow==10.1.0
PyYAML==6.0
python-dotenv==1.0.0
gunicorn==20.1.0
django-filter==21.1
psycopg2-binary==2.9.9
django-import-export==3.3.4
fpdf2==2.7.7